listener/requirements.txt
nameko==2.14.0
pytest==4.5.0
coverage==4.5.3
flake8==4.0.1
flake8-annotations==2.7.0
websocket-client==1.2.1